feat: split menu into two - mobile and desktop
